# API/openapi.yaml
openapi: 3.0.3
info:
  title: UnlikeOtherAuthenticator API
  version: 1.0.0
  description: |
    Centralized OAuth / authentication service. Provides user auth flows,
    token issuance, organisation/team management, and admin operations.

    ## Authentication Schemes

    | Scheme | Header | Format | Used By |
    |--------|--------|--------|---------|
    | Domain Hash Bearer | `Authorization` | `Bearer SHA256(normalizedDomain+clientSecret)` | Domain-scoped server-to-server calls |
    | Admin Bearer | `Authorization` | `Bearer <HS256 JWT>` | Internal admin endpoints |
    | Access Token | `X-UOA-Access-Token` | `Bearer <JWT>` or raw JWT | User-context org endpoints |
    | Config JWT | `X-UOA-Config-JWT` | RS256 JWT | Email send endpoint |

    ## Config URL
    Most endpoints require a `config_url` query parameter pointing to a signed
    RS256 config JWT hosted by the integrator. The server fetches and verifies
    this JWT on every request.
servers:
  - url: https://auth.example.com
    description: Production
security: []
...
components:
  securitySchemes:
    domainHashBearer:
      type: http
      scheme: bearer
      description: |
        `Bearer SHA256(normalizedDomain + clientSecret)` — uses the per-domain client secret, not the shared secret.
    adminBearer:
      type: http
      scheme: bearer
      description: |
        HS256 JWT signed with `ADMIN_ACCESS_TOKEN_SECRET`. Claims must include
        `role: "superuser"` and `domain` matching `ADMIN_AUTH_DOMAIN`.
    accessToken:
      type: apiKey
      in: header
      name: X-UOA-Access-Token
      description: |
        UOA access token (HS256 JWT). May be prefixed with `Bearer ` or sent raw.
    configJwt:
      type: apiKey
      in: header
      name: X-UOA-Config-JWT
      description: RS256 config JWT signed with the domain's private key.
    billingAppKey:
      type: apiKey
      in: header
      name: X-UOA-App-Key
      description: One revocable UOA app key dedicated to the exact calling product/deployment and either entitlement or customer-lifecycle endpoints.
    billingActor:
      type: apiKey
      in: header
      name: X-UOA-Actor
      description: Short-lived RS256 actor assertion bound to the exact product app key and UOA user/org/team.
    stripeSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
      description: Stripe endpoint signature verified over the exact raw request body with a separate webhook secret.
  parameters:
...
    domainQuery:
      name: domain
      in: query
      required: true
      description: Normalised domain (lowercase, trimmed).
      schema:
        type: string
        example: acme.example.com
...
    avatarStyleQuery:
      name: style
      in: query
      required: false
      description: |
        Overrides the generated-avatar style for this call. Only affects generated
        avatars — uploaded and proxied provider images are served as stored.
      schema:
        type: string
        enum: [tiles, waves, rings, mono]
    avatarSizeQuery:
      name: size
      in: query
      required: false
      description: |
        Rendered width/height of a generated SVG, clamped to 16-512 (default 128).
        The viewBox is constant; ignored for raster images.
      schema:
        type: integer
        minimum: 16
        maximum: 512
        default: 128
    avatarConfigUrlQuery:
      name: config_url
      in: query
      required: false
      description: |
        Optional HTTPS URL to the domain's signed config JWT. When supplied it is fetched
        and verified, and its `avatars.default_style` claim is used when `style` is absent.
        Its `domain` claim must match the authenticated domain.
      schema:
        type: string
        format: uri
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Generic public error message.
        code:
          type: string
          description: Machine-readable error code (only when DEBUG_ENABLED).
        summary:
          type: string
        details:
          type: array
          items:
            type: string
        hints:
          type: array
          items:
            type: string
...
    AvatarUploadResponse:
      type: object
      required: [ok, avatar]
      properties:
        ok:
          type: boolean
          enum: [true]
        avatar:
          type: object
          required: [source, content_type, size_bytes, updated_at]
          properties:
            source:
              type: string
              enum: [uploaded]
            content_type:
              type: string
              enum: [image/png, image/jpeg, image/webp]
              description: The magic-byte sniffed type that was stored.
            size_bytes:
              type: integer
            updated_at:
              type: string
              format: date-time
    AvatarUploadRequest:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
          description: |
            PNG, JPEG or WebP, at most 1 MiB. The stored type is decided by magic-byte
            sniffing, not the supplied mimetype; SVG and other non-raster uploads are
            rejected with the generic error envelope.
...
paths:
...
  /avatar/me:
    get:
      tags: [Avatars]
      summary: Get the access-token subject's avatar image
      operationId: getMyAvatar
      description: |
        Same resolution pipeline as the domain route, for the caller's own user. The access
        token's `domain` claim must equal `domain`; the acting identity is always the token
        subject.
      security:
        - domainHashBearer: []
          accessToken: []
      parameters:
        - $ref: '#/components/parameters/domainQuery'
        - $ref: '#/components/parameters/avatarConfigUrlQuery'
        - $ref: '#/components/parameters/avatarStyleQuery'
        - $ref: '#/components/parameters/avatarSizeQuery'
      responses:
        '200':
          description: Avatar image bytes.
          headers:
            X-UOA-Avatar-Source:
              schema:
                type: string
                enum: [uploaded, provider, generated]
          content:
            image/svg+xml:
              schema:
                type: string
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        '304':
          description: Not modified (matching `If-None-Match`).
        '401':
          description: Missing or invalid domain hash auth or access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Access token domain does not match `domain`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Avatars]
      summary: Set the caller's own avatar
      operationId: putMyAvatar
      description: |
        Uploads the access-token subject's own avatar. The acting identity is always the
        token subject. Rate-limited per domain + user.
      security:
        - domainHashBearer: []
          accessToken: []
      parameters:
        - $ref: '#/components/parameters/domainQuery'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AvatarUploadRequest'
      responses:
        '200':
          description: Stored avatar metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvatarUploadResponse'
        '400':
          description: Generic error — not a supported raster image.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Missing or invalid domain hash auth or access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: Upload exceeds the 1 MiB cap.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Avatars]
      summary: Clear the caller's own uploaded avatar
      operationId: deleteMyAvatar
      description: |
        Idempotent. Resolution falls back to the provider URL or the generated image.
        Rate-limited per domain + user.
      security:
        - domainHashBearer: []
          accessToken: []
      parameters:
        - $ref: '#/components/parameters/domainQuery'
      responses:
        '200':
          description: Cleared.
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
                    enum: [true]
        '401':
          description: Missing or invalid domain hash auth or access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...